+++ comp2013/src/main/java/module-info.java
@@ -2,4 +2,7 @@ module comp2013 {
     opens comp2013;
     requires java.desktop;
     requires jlayer;
+    requires javafx.graphics;
+    requires javafx.controls;
+    requires javafx.fxml;
 }