+++ docs/changelog.md
@@ -26,3 +26,58 @@
 - Reorganised class files.
 - Changed some file locations in code to fit with the new Maven Build Method.
 - Added some new issues to Git.
+### 29/11/23
+- Wrote up plan for MVC in DevNotes.
+- Added Interface Classes for MVC.
+- Implemented Methods for MVC interfaces.
+- Implemented Constants in IController Interface.
+- Added Class files for MVC.
+- Removed Redundant Main Class.
+### 30/11/23
+- Implemented Getters and Setters in the Model.
+- Added some private variables in Model.
+- Impmemented part of Initialise method in Model.
+- Started to implement InitialiseDisplay method in view.
+- Changed variable names so that they follow the coding convetntions. 
+
+### 6/12/23
+- Changed comments on SnakeView from JavaDoc to Regular Comments, as JavaDoc is inherited from the Interface Class.
+- Changed comments on SnakeController from JavaDoc to Regular Comments, see previous.
+- Added two methods to the model interface; getAlive and setAlive, to get and set the liveliness status of the snake, to be used for things like ending the game when the snake dies.
+- Added a new variable to SnakeModel, m_Alive.
+- Implemented methods in SnakeModel, incrementLength, getAlive, setAlive.
+- Reworked SnakeView initialiseDisplay to respect MVC, view now has no interaction with model.
+- Added constructors to View, Controller and Model.
+- Added a new Class, SnakeMain, that intialised all of the MVC.
+- Started to convert the program to JavaFX. 
+- Set up JavaFX.
+- Fixed an issue where the controller was not initialised in the View beacuse of a issue with how JavaFX handles its instance of View.
+- Added an Icon to the JavaFx.
+- Implemented FXML.
+- Updated Maven POM with dependencies.
+- Updated module-info.
+- Linked FXML to the Controller and View.
+
+### 7/12/23
+- Fixed issue with controller.
+- Added image to screen.
+- Added Score Label to Screen.
+- Create a new Class SnakeObject, from the SnakeObject class in MyFrame.
+- Rename interface Movable.
+- Add new functions to interface Movable.
+- Create a new class SnakeBody that stores information about the snake body parts.
+- Added a new method buildSnake in SnakeView that builds the initial snake.
+- Added a new method in Controller addSegment that appends a new segment to the list in SnakeObject.
+- Add new method to IController handleKeyPress.
+- Add method to View, refreshSnake, which rebuilds the entire snake to the new loaction.
+- Added constants for WSAD and Arrow Key, Key codes
+- Added method handleKeyPress to Controller to handle what happens when a key is pressed.
+- Added code to view to start the infinite loop that refreshes the screen.
+- Reimplemented eatBody as getSelfCollide in Controller.
+- Rename getSelfCollide to checkSelfCollide.
+- Removed getSnakeMoving from controller as it is not needed.
+- Added handleGameover function to Controller.
+- Added 3 new Images, snake-head-"left", "up", "down".
+- Added method to change the direction the image is facing.
+- Update moveSnake to add snake rotation functionality from view.
+- Fixed bug with controller not being set view correctly.